Get utils.cpp to compile without warnings under gcc so that it can be used in my wxWidgets experiment.
git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2374 9d0050d7-00b0-40f9-871c-8e02cbd5794f

+++ trunk/logo/src/utils.cpp
@@ -19,7 +19,8 @@
  *
  */
 
-#include "allwind.h"
+#include <windows.h>
+#include <stdio.h>
 
 // the name of the FMSLogo registry key under HKCU
 const char FMSLOGO_REGISTRY_KEY_NAME[] = "Software\\FMSLogo";
@@ -378,7 +379,7 @@ SetConfigurationFont(
         {FONTPROPERTY_PitchAndFamily,  LogFont.lfPitchAndFamily},
     };
 
-    for (int i = 0; i < sizeof(properties) / sizeof(*properties); i++)
+    for (size_t i = 0; i < sizeof(properties) / sizeof(*properties); i++)
     {
         strcpy(relativeName, properties[i].Name);
         SetConfigurationInt(fullyQualifiedName, properties[i].Value);
